+++ src/Root.tsx
@@ -1,7 +1,7 @@
 import { motion, useAnimationControls } from "framer-motion";
 import React, { useEffect, useState } from "react";
 
-import { FirstLoadView } from "./common/FirstLoadView/FirstLoadView";
+import { FirstLoadView } from "./firstLoadView/FirstLoadView";
 import { Project } from "./domain/type";
 import { MainContent } from "./MainContent";
 import { styled } from "./stitches.config";
@@ -10,7 +10,7 @@ type Props = {
   project: Project;
 };
 
-const RootContainer: React.FC<Props> = ({ project }) => {
+export const RootContainer: React.FC<Props> = ({ project }) => {
   const [isFirstLoadCompoleted, setIsFirstLoadCompleted] = useState(false);
   const mainPartControls = useAnimationControls();
   const onFirstLoadComplete = () => {
@@ -31,7 +31,7 @@ const RootContainer: React.FC<Props> = ({ project }) => {
           <FirstLoadView receiverName={project.receiver_name} onLoadComplete={onFirstLoadComplete} />
         </FirstLoadPart>
       ) : (
-        <MainPart
+        <motion.div
           animate={mainPartControls}
           style={{ opacity: 0, overflow: "hidden" }}
           transition={{
@@ -40,7 +40,7 @@ const RootContainer: React.FC<Props> = ({ project }) => {
           }}
         >
           <MainContent />
-        </MainPart>
+        </motion.div>
       )}
     </>
   );
@@ -54,11 +54,3 @@ const FirstLoadPart = styled("div", {
   width: "100vw",
   height: "100vh",
 });
-export const MainPart = motion(styled("div", {}));
-
-export default RootContainer;
-
-const Root: React.FC<Props> = ({ project }) => {
-  const senderNameList = project.text_messages.items.map((message) => message.sender_name);
-  return <div></div>;
-};
