+++ src/music/Music.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import { styled } from "./stitches.config";
-import { Base } from "./utils/ui";
+import { styled } from "../stitches.config";
+import { Base } from "../utils/ui";
 
 export const Music: React.FC = () => {
   return (
